Factor out treatment outcome component for usee in updating treatment outcomes
